Integrated sidebar into students page
