+++ thi-app/app/index.tsx
@@ -1,6 +1,6 @@
+import Login from '@/components/login';
 import React from 'react';
 import { View } from 'react-native';
-import Login from '../components/Login';
 
 export default function LoginPage() {
   return (
